+++ src/datasets/domain/useCases/GetDatasetVersionsSummaries.ts
@@ -1,8 +1,8 @@
 import { UseCase } from '../../../core/domain/useCases/UseCase'
-import { DatasetVersionSummaryInfo } from '../models/DatasetVersionSummaryInfo'
+import { DatasetVersionSummarySubset } from '../models/DatasetVersionSummaryInfo'
 import { IDatasetsRepository } from '../repositories/IDatasetsRepository'
 
-export class GetDatasetVersionsSummaries implements UseCase<DatasetVersionSummaryInfo[]> {
+export class GetDatasetVersionsSummaries implements UseCase<DatasetVersionSummarySubset> {
   private datasetsRepository: IDatasetsRepository
 
   constructor(datasetsRepository: IDatasetsRepository) {
@@ -14,9 +14,15 @@ export class GetDatasetVersionsSummaries implements UseCase<DatasetVersionSummar
    * Draft versions will only be available to users who have permission to view unpublished drafts.
    *
    * @param {number | string} [datasetId] - The dataset identifier, which can be a string (for persistent identifiers), or a number (for numeric identifiers).
-   * @returns {Promise<DatasetVersionSummaryInfo[]>} - An array of DatasetVersionSummaryInfo.
+   * @param {number} [limit] - Limit for pagination (optional).
+   * @param {number} [offset] - Offset for pagination (optional).
+   * @returns {Promise<DatasetVersionSummarySubset>} - A DatasetVersionSummarySubset containing the summaries and total count.
    */
-  async execute(datasetId: number | string): Promise<DatasetVersionSummaryInfo[]> {
-    return await this.datasetsRepository.getDatasetVersionsSummaries(datasetId)
+  async execute(
+    datasetId: number | string,
+    limit?: number,
+    offset?: number
+  ): Promise<DatasetVersionSummarySubset> {
+    return await this.datasetsRepository.getDatasetVersionsSummaries(datasetId, limit, offset)
   }
 }